+++ laktory/models/databricks/workspacefile.py
@@ -30,16 +30,8 @@ class WorkspaceFile(BaseModel, BaseResource):
 
     @property
     def filename(self) -> str:
-        """File file name"""
+        """File filename"""
         return os.path.basename(self.source)
-
-    @property
-    def key(self) -> str:
-        """File resource key"""
-        key = os.path.splitext(self.path)[0].replace("/", "-")
-        if key.startswith("-"):
-            key = key[1:]
-        return key
 
     @model_validator(mode="after")
     def default_path(self) -> Any:
@@ -56,6 +48,16 @@ class WorkspaceFile(BaseModel, BaseResource):
                 )
 
         return self
+
+    @property
+    def resource_key(self) -> str:
+        """File resource key"""
+        key = os.path.splitext(self.path)[0].replace("/", "-")
+        if key.startswith("-"):
+            key = key[1:]
+        return key
+
+    # resource_key.__doc__ = super().resource_key.__doc__
 
     # ----------------------------------------------------------------------- #
     # Resources Engine Methods                                                #