+++ laktory/models/sql/schema.py
@@ -93,6 +93,10 @@ class Schema(BaseModel, BaseResource):
     # ----------------------------------------------------------------------- #
 
     @property
+    def resource_key(self) -> str:
+        return self.full_name
+
+    @property
     def pulumi_excludes(self) -> list[str]:
         return ["tables", "volumes", "grants"]